+++ src/template/src/client/App.tsx
@@ -5,7 +5,10 @@ import BeerDisplayer from "./components/BeerDisplayer";
 import LoadingIndicator from "./components/LoadingIndicator";
 import Modal from "./components/Modal";
 import { Beer } from "./types";
-/* NEW LINE */
+
+/**
+ * Application.
+ */
 const App: React.SFC<{}> = () => {
   const [beers, setBeers] = React.useState<[] | Beer[]>([]);
   const [isModalOpen, setIsModalOpen] = React.useState(false);