+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Comments in template files [#22](https://github.com/alexlee-dev/create-mern-application/issues/22)
+
 ### Changed
 
 ### Removed