+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+\# Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Comments in template files [#22](https://github.com/alexlee-dev/create-mern-application/issues/22)
+- Comments in template files - [#22](https://github.com/alexlee-dev/create-mern-application/issues/22)
+- README `Built With` section - [#9](https://github.com/alexlee-dev/create-mern-application/issues/9)
 
 ### Changed